Move ScrollFlagArea to panels.scrollflag.py
